% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSig.R
\name{computeCoExp}
\alias{computeCoExp}
\title{Compute co-expression networks given seurat object for each sample and
Perform unsupervised HOSVD on the combined 3D tensor co-expression network}
\usage{
computeCoExp(
  obj,
  prefix = ".",
  sampleid = "Sample_ID",
  recompute_glasso = TRUE,
  recompute_HOSVD = TRUE,
  ReComputeVarGene = FALSE,
  mean_thres = 0.1
)
}
\arguments{
\item{prefix}{folder for raw output (automatic create raw folder)}

\item{sampleid}{column name in seurat containing sample info}

\item{recompute_glasso}{recompute the graphical lasso co-expression network}

\item{recompute_HOSVD}{recompute HOSVD for the co-expression network}

\item{ReComputeVarGene}{recompute the variable genes}

\item{mean_thres}{minimal threshold for mean expression of variable genes}

\item{return}{Decomposed tensor matrices , list of sample names , and list of gene names}
}
\description{
Compute co-expression networks given seurat object for each sample and
Perform unsupervised HOSVD on the combined 3D tensor co-expression network
}
