% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSig.R
\name{featureAnalysis}
\alias{featureAnalysis}
\title{Extract the tensor component (TC) that correlated with the feature of interest,
and construct the network}
\usage{
featureAnalysis(
  prefix,
  cp_decomp,
  t_coexp,
  sampleNames,
  varFeatures,
  meta,
  featureOfInterest,
  featureOfCategory,
  cor_method = "spearman",
  zthres = 2
)
}
\arguments{
\item{prefix}{working directory prefix}

\item{cp_decomp}{decomposed matrices from HOSVD}

\item{t_coexp}{3D tensor matrix for stacked co-expression networks for all samples}

\item{sampleNames}{names of the samples}

\item{varFeatures}{gene names}

\item{meta}{metadata for samples}

\item{featureOfInterest}{column name of the feature of interest in the metadata assuming the features are levels or numerical values}

\item{featureOfCategory}{column name of the grouping in the metadata: assuming the features are levels}

\item{cor_method}{correlation method}

\item{zthres}{threshold for z-score of loading values to identify significant driving genes per TC}
}
\description{
Extract the tensor component (TC) that correlated with the feature of interest,
and construct the network
}
