% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSig.R
\name{initializeMultiCores}
\alias{initializeMultiCores}
\title{Setting multi-threading}
\usage{
initializeMultiCores(cores = 10)
}
\arguments{
\item{cores:}{Number of cores to use, not to overload your computer}
}
\description{
Setting multi-threading
}
