% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSig.R
\name{selectGenesCo}
\alias{selectGenesCo}
\title{Extract the genes with highest loadings to TCs}
\usage{
selectGenesCo(cp_decomp, selectComp, varFeatures, zthres = 2)
}
\arguments{
\item{cp_decomp:}{decomposed matrices from HOSVD}

\item{varFeatures:}{gene names}

\item{zthres:}{threshold for z-score of loading values to identify significant driving genes per TC}
}
\value{
A list of driving genes
}
\description{
Extract the genes with highest loadings to TCs
}
